feat: 우승자 출력 추가 및 출력 메소드명 변경
